Fix job `error_message` (#48)
* Adding cython

* Inpecting job results for getting error details.

When a job fails, requesting the result to the API returns the error
details.

* Adding error details

* Handling several errors mixed with success

* Fixing style

* Reverting .gitignore

* Avoiding consuming the job unless needed.

Some jobs can be retrieved only once. A successful query to the job's
complete state involves no further access. I call this effect "consuming
a job".

This commit rewrites the implementation of IBMQ job to isolating the
consumption into two methods: `_wait_for_job` and `_get_job` which are
semantically identical but the first will block until completion.

For waiting until completion but without consuming the job, there is a
new method `_wait_for_completion`. Most of the rewriting delves with
using this new method instead of the removed `_wait_for_result`,
allowing finer control and new actions between completion and getting
the results.

This enables consuming the job only when it is DONE and after calling
`result()`, or after calling `error_message()` only if the status is
ERROR.

* Linting

* Review improvements

* Cache result to allow additional calls.

* Clarifying consuming behaviours of error_message and result methods