Move CONTRIBUTING.rst to top level (#149)
* Move CONTRIBUTING.rst to top level

The CONTRIBUTING.md doc contains a lot of useful information about the
not only contributing to the project but also some useful information
about project policy and conventions. But it's currently in a hidden
directory so that github ui will show it on a help menu. So unless you
know to click on that help menu or that it's hiding in the .github
directory you'll never see it. This commit addresses this by moving the
file to the repo root so it's more obvious where it is. This also
converts it to be in the markdown format to be consistent with the other
repos (qiskit-terra will be also converted in Qiskit/qiskit-terra#2359).

Mirror of Qiskit/qiskit-terra#2359 Qiskit/qiskit-aer#229
Qiskit/qiskit-ignis#216 and Qiskit/qiskit-aqua#556

* Move code of conduct and convert changelog to md