Skip online tests if no credentials are found
Right now if there are no credentials set in the environment all the
online tests fail. This is not desirable because you shouldn't need
credentials to run the tests successfully. This commit changes the
raised exception to be a skip exception to make tests that require
credentials skip instead of fail when they can't be run.