Set up CI with Azure Pipelines to replace appveyor (#203)
* Set up CI with Azure Pipelines to replace appveyor

This commit adds a yaml config to start using azure pipelines instead of using appveyor for our windows testing. For our purposes it's mostly the same except that azure is faster and provides much more capacity. [skip ci]

* Fix typo

* Remove appveyor

* Skip online tests if no credentials are found

Right now if there are no credentials set in the environment all the
online tests fail. This is not desirable because you shouldn't need
credentials to run the tests successfully. This commit changes the
raised exception to be a skip exception to make tests that require
credentials skip instead of fail when they can't be run.

* Fix lint

* Revert "Skip online tests if no credentials are found"

This reverts commit c3b8e57dfee86b164e1f4e961d2c5c84039e8841.

* Revert "Fix lint"

This reverts commit c5d49f392769fad1329a681fd447adb8e2ab7ef6.

* Set env var to skip online tests

* Conditionally set online tests skip

* Fix copy and paste issue

* Create duplicated job definitions for non-PR builds

* Fix typo

* Fix another typo

* Add missing paren

* Update azure pipelines script

Update the script in order to use only one job temporarily, and attempt
to avoid double execution of the jobs using a condition.