Add pypi automation for releases
As part of streamlining the release process, this commit adds automation
for publishing to pypi. The release process is straightforward as we just
need to build an sdist and a python3 wheel and then upload both to pypi.
This takes another manual step out of the release process and all that
is necessary is to push a tag (or use github's ui for creating a tag
and/or release). This new travis job will handle the whell and sdist
along with the pypi upload.