Prepare for Bugfix release (#314)
* don't use newer asyncio (#312)

* Actually pin requirement and bump version for release

This commit pins the nest_asyncio in the actual requirements list and
also bumps versions and prepares the changelog for release.

* Add pypi automation for releases (#278)

As part of streamlining the release process, this commit adds automation
for publishing to pypi. The release process is straightforward as we just
need to build an sdist and a python3 wheel and then upload both to pypi.
This takes another manual step out of the release process and all that
is necessary is to push a tag (or use github's ui for creating a tag
and/or release). This new travis job will handle the whell and sdist
along with the pypi upload.