Set language for travis wheel publishing job (#316)
Currently the wheel publishing job fails early when you run pip install
twine. This is because the job doesn't have a lanaguage set and travis
defaults to ruby. The ruby language jobs don't have the proper
permissions set up for running pip without extra permissions (ie sudo).
To fix the failure for future releases this commit explicitly sets the
job to be python so that we can install twine and actually build the
sdist and wheel.