Don't allow zipped eggs
Zipped eggs can cause issues with namespaced packages. We've actually
seen instances of this in Qiskit/qiskit-ignis#164 and
Qiskit/qiskit-ignis#175. The current recommendation from the python
packaging community is to set zip_safe to false for namespaced packages.
This commit does this for the project to avoid any issues caused by this
in the future.

[1] https://github.com/pypa/sample-namespace-packages#current-status