Revert "don't build wheels (#454)" (#455)
* Revert "don't build wheels (#454)"

Wheels are useful because they're more efficient they're not just
about binaries. Wheels are a distribution format and are just unzipped
when installed, while sdists require running the setup.py and building
the package to be installed locally. While wheels are very useful for
compiled python extensions because it lets us distribute pre-compiled
binaries the benefits extend beyond just that. We should still build
and publish a wheel even if it's just python code. This is why there is
a separate definition for both universal and pure python wheels. [1]

This reverts commit 8d0d8df88f681d29cb00e596c2e05a39024a12c5.

[1] https://packaging.python.org/guides/distributing-packages-using-setuptools/#packaging-your-project

* Undo ibmqjob change revert