Pin and unify lint job versions used across elements (#496)
* Pin and unify lint job versions used across elements

In an effort to unify the versions of pylint run across all the elements
this commit sets an explicit pin via a constraints file on the version
we run in CI. This ensures consistent lint job behavior between all the
qiskit elements. Also pinning pylint versions is good practice when
running in CI because every pylint and/or astroid release typically
introduces behavior changes. To ensure we have a stable base to build
off we should ensure we have a fixed version of pylint (and astroid
which has a hard coupling for behavior to) and that we bump it
intentionally across all the elements.

The python version used for the lint job in ci is also updated to use
the same version as all other elements, 3.7. Pylint's behavior is
dependent on python version (mostly do to stdlib differences between
python versions) so to ensure consistent behavior between elements we
need to ensure that everything is running lint with the same python
version.

Related To: Qiskit/qiskit-terra#3629

* Fix issues with new versions

Co-authored-by: Jessie Yu <jessieyu@us.ibm.com>