Disable jsonschema validation by default
Running jsonschema validation is exceedingly slow, especially for large
payloads. It also provides no functional benefit since the api service
will validate the input already. If a structurally invalid request is
sent to the api it will reject it with a 400 response. This commit adds
a new option to the run method on the backend() to disable running
validation by default. This option is then used in testing so we run
validation in tests where speed doesn't matter.