Define custom json encoder for qobj payloads (#556)
The qobj payload has a couple of quirks related to how qobj objects can
be constructed (and will be constructed after Qiskit/qiskit-terra#3383
merges) that aren't handled by the default json encoder. The first is
that complex numbers are converted to the form "(real, imaginary)", and
the second is that numpy arrays are sometimes part of the payload.
Instead of relying on a layer outside of the provider coercing these
objects to the expected format this commit changes how we push qobj
payloads to the iqx api so that we run it through a json encoder that
understands how to do this. After this commit qiskit-ibmq-provider owns
it's own payload format and the generation of the expected json payload
format instead of relying on an external library to handle it for us.

Related to #553

Co-authored-by: Jessie Yu <jessieyu@us.ibm.com>