Handled parameterized circuits
After Qiskit/qiskit-terra#3383 the conversion of types outside the
standard set of types understood by the standard json encoder is the
responsibility of the ibmq provider. However, there was one type missing
from the custom encoder class added to handle this, ParameterExpression.
For bound parameterized circuit the ParameterExpression class will be
castable to a float to get the bound numeric value that should be sent
on the wire with the job. However, the encoder doesn't have handling for
this. Qiskit/qiskit-terra#3959 is adding conversion for the to_dict()
qobj method so that floats are output, but this a stop gap workaround
while a better solution is being developed. We should have handling
in qiskit-ibmq-provider if/when ParameterExpression objects get passed
in via qobj object.