Add reno for release documentation and update CONTRIBUTING (#679)
* Add reno for release documentation and update CONTRIBUTING

Moving forward qiskit-ibmq-provider will be using reno for release
documentation. This commit starts using reno for release notes by
converting the existing changelog entry into a reno release note and
adding reno to the sphinx configuration to verify release notes as part
of documentation builds. As part of this the CONTRIBUTING guide was
updated to reflect this change, at the same time the guide was out of
date and mostly duplicated with the project wide documentation. In the
near future the reno and release pieces added here will be moved to the
cross-project documentation too, but until then this adds a local copy.

* Fix whitespace

* Update installing from source link

* Add release note for new changelog entry

* Apply suggestions from code review

Co-authored-by: Jessie Yu <jessieyu@us.ibm.com>

Co-authored-by: Jessie Yu <jessieyu@us.ibm.com>