Fix module name in ibmq apidoc index (#740)
* Fix module name in ibmq apidoc index

As part of the combined documentation build sphinx emits a warning:
"WARNING: duplicate object description of qiskit, other instance in
apidoc/ibmq-provider, use :noindex: for one of them". This is caused by
the module name for the ibmq provider index incorrectly saying this
document is for the module qiskit; instead it should be
qiskit.providers.ibmq which is actually the module for the index. This
commit corrects the oversignt.

* Remove module line