Pass hub/group/project to POST /experiments (#796)
* Pass hub/group/project to POST /experiments

In an upcoming change in the `POST /experiments` API, experiments
are going to require storing the hub_id/group_id/project_id. This
change provides those values when creating an experiment.

There are questions inline about how to do this, i.e. should the
Experiment object hold those fields or should we just get them
from the credentials on the account provider as needed (which is done
here).

There is also not a great way to test this right now since the server
is not yet providing the fields in the response which we're passing in
on the request so a TODO is left for that.

Partial #794

* Store hub/group/project on Experiment

This sets the immuatable hub/group/project attributes on
the Experiment instance and uses those when uploading (creating)
the experiment in the service.

* Assert experiment is created with hub/group/project

* Set hub/group/project when retrieving experiments

* Fix format string error