Expose Experiment visibility and owner attributes
This adds an enum for the Experiment.visibility field
and allows it to be set when creating the experiment
and change it when updating the experiment. Since it can
be updated the Experiment object defaults visibility to
None so that the server can set the default (private).
I mostly copied how that works from AnalysisResult.quality.

The experiment owner is set by the server and cannot be
updated.

While updating the Experiment.__repr__ method I noticed that
it was missing hub/group/project so I also added those in.

Closes #825