Use verify_headers.py from terra instead of pylintfileheader
The pylintfileheader pylint plugin to verify copyright headers
conform to the project guidelines is problematic for a number of reasons.
The first is it has proven to be unstable there have been several
instances of both users and CI failing on header files that
do not have any issues. This is pretty common with pylint in general
since it's very sensitive to differences in environment. Additionally,
it was using a regex to try and match the entire header string. This
provided no practical debugability when it fails (especially when the
failure is unexpected) because we end up with only a match error.
This commit addresses these issues by removing the use of the regex
based pylint plugin and adding a small script to verify the header in
every qiskit file and provide useful debugging if there is an issue with
a header in a python file.

This is borrowed from the same script in qiskit-terra, you can see the
full history of it at the point of copying it:
https://github.com/Qiskit/qiskit-terra/blob/5d82781e2467e6663462f939884b80345ff1b2c9/tools/verify_headers.py

Fixes #828