Support filtering experiments by public visibility
This adds an `is_public` keyword filter when listing experiments.
By default the behavior is the same as today, there is no `visibility`
filter applied. But if set to True then only public experiments will
be returned. If set to False then only non-public experiments will be
returned.

This also fixes a typo in a previous release note.

Closes #848