Support filtering experiments by public share_level (#858)
* Support filtering experiments by public visibility

This adds an `is_public` keyword filter when listing experiments.
By default the behavior is the same as today, there is no `visibility`
filter applied. But if set to True then only public experiments will
be returned. If set to False then only non-public experiments will be
returned.

This also fixes a typo in a previous release note.

Closes #848

* Change is_public to exclude_public/public_only kwargs

This splits the possible triple value is_public kwarg into
distinct exclude_public and public_only kwargs with the following
API behavior:

| exclude_public | public_only | result                         |
| -------------- | ----------- | ------------------------------ |
| False          | False       | no visibility filter (default) |
| True           | False       | visibility=!public             |
| False          | True        | visibility=public              |
| True           | True        | raise error                    |

* Add tests for exclude_public/public_only filtering

* Add newline to release note

* Disable unexpected-keyword-arg in test

I don't know why the E1123 error is showing up in CI, it doesn't
fail locally and there is definitely a `share_level` kwarg on the
Experiment class. Disable for now.

* split tests

* fix lint

Co-authored-by: jessieyu <jessieyu@us.ibm.com>