Handle asyncio breaking change
In python 3.7 the class method asyncio.Task.all_tasks() was deprecated
and replaced with the function asyncio.all_tasks(). In python 3.9 the
asyncio.Task.all_tasks() class method was removed which is resulting on
test failures with python 3.9. Still we still support python 3.6 too to
fix the test failure we need to use asyncio.Task.all_tasks() with python
3.6 and asyncio.all_tasks() everywhere else. When support for python 3.6
is dropped we can remove this version check and rely on just
asyncio.all_tasks() everywhere.