Add support for Python 3.9 (#862)
* Add support for Python 3.9

Python 3.9.0 was released on 10-05-2020, this commits marks the support
of Python 3.9 in qiskit-ibmq-provider. It adds the supported python
version in the package metadata and updates the CI configuration to run
test jobs on Python 3.9.

* Add release notes

* Use terra from git for py39 testing

* Fix typo

* Handle asyncio breaking change

In python 3.7 the class method asyncio.Task.all_tasks() was deprecated
and replaced with the function asyncio.all_tasks(). In python 3.9 the
asyncio.Task.all_tasks() class method was removed which is resulting on
test failures with python 3.9. Still we still support python 3.6 too to
fix the test failure we need to use asyncio.Task.all_tasks() with python
3.6 and asyncio.all_tasks() everywhere else. When support for python 3.6
is dropped we can remove this version check and rely on just
asyncio.all_tasks() everywhere.

* Change if condition to < 3.9

* Pass in event loop

Co-authored-by: Jessie Yu <jessieyu@us.ibm.com>