Add Good/Bad ResultQuality enum values
The Computer/Human analysis result `quality` values are
being deprecated and replaced with Good/Bad `quality` values.
The `verified` field indicates if a human verified the quality,
so `Human Bad` maps to `quality='Bad'` and `verified=True` while
`Computer Good` maps to `quality='Good'` and `verified=False`.
The server handles the mapping of the old deprecated `quality`
values when an analysis result is created or updated.

Existing data has already been migrated so when getting analysis
results back they should have `quality` of `Bad`, `No Information`
or `Good` which means we need to add the new `Bad` and `Good`
values to the `ResultQuality` enum.

Part of #902