Avoid use of __qiskit_version__ for client version header (#910)
* Avoid use of __qiskit_version__ for client version header

The ibmq provider was using the __qiskit_version__ attribute to try and
determine the various qiskit versions for inclusion in a client version
header. At first glance this would seem like the correct way to handle
this, however it actually is problematic. This is because
__qiskit_version__ imports all the installed qiskit packages (including
things completely unrelated to the use-case for the client version
header) to provide a single end-user entrypoint for debugging to get all
the qiskit package versions. This is undesirable because the
qiskit-ibmq-provider usage of the __qiskit_version__ attribute happens
on import, this means that we'll be unconditionally importing the entire
set of qiskit packages even if only terra and the qiskit provider are
used. This has the negative effect of both significantly increasing
import time (as things like aqua have very heavy dependencies and a slow
import tree) but also by potentially causing undesireable deprecation
warnings to be emitted if there are things in the import path that are
deprecated (like aqua). To fix this issue this commit switches away from
using the __qiskit_version__ attribute and queries the qiskit
metapackage version directly, and if it's not available then the ibmq
provider and terra versions are set in the header. This avoids the
import related issues from using __qiskit_version__ while not changing
the functionality (because the other non-terra requirements in the
header should have no impact on the api behavior).

The client version used in the header should honestly be independent of
anything outside of the qiskit-ibmq-provider as it's the only
interaction point between the iqx api and qiskit (and things I believe
this was added for like the basis gate change can be handled by the
minimum version of terra in the requirements). Doing it this way just adds
an unnecessary dependency and complexity to version negotiation. But it's
too late to change the iqx api design at this point.

* update header format

Co-authored-by: jessieyu <jessieyu@us.ibm.com>
(cherry picked from commit f5c3b5003e2c71c3befba7f410cbcba4008f6771)