Make AnalysisResult.fit optional
The `fit` attribute is now optional in the service and
if specified it does not need `value` or `variance` values,
it's just a free-form dict and defaults to an empty dict.
As such, the special `Fit` class is removed here and tests
are updated.

To get tests to pass, `test_analysis_results_quality` is
skipped until issue #902 is fixed.

Also while in here, the temporary `hello_world.svg` plot
file is removed on cleanup of the test case.

Closes #916