[Stable] Release 0.12.3 (#920)
* skip aer partial result tests (#911)

(cherry picked from commit bebc5cdf8ea271444ecdc4a57c4224370fdc452d)

* Reduce v2 provider warning (#912)

* warn only once

* warn only once

* fix lint

(cherry picked from commit 176cc07f556c5c026a15b1aec8860bdd003733e9)

* Add Good/Bad ResultQuality enum values (#915)

The Computer/Human analysis result `quality` values are
being deprecated and replaced with Good/Bad `quality` values.
The `verified` field indicates if a human verified the quality,
so `Human Bad` maps to `quality='Bad'` and `verified=True` while
`Computer Good` maps to `quality='Good'` and `verified=False`.
The server handles the mapping of the old deprecated `quality`
values when an analysis result is created or updated.

Existing data has already been migrated so when getting analysis
results back they should have `quality` of `Bad`, `No Information`
or `Good` which means we need to add the new `Bad` and `Good`
values to the `ResultQuality` enum.

Part of #902

(cherry picked from commit fe5a07c7f8983128425b094f0d28daa01f9addec)

* Avoid use of __qiskit_version__ for client version header (#910)

* Avoid use of __qiskit_version__ for client version header

The ibmq provider was using the __qiskit_version__ attribute to try and
determine the various qiskit versions for inclusion in a client version
header. At first glance this would seem like the correct way to handle
this, however it actually is problematic. This is because
__qiskit_version__ imports all the installed qiskit packages (including
things completely unrelated to the use-case for the client version
header) to provide a single end-user entrypoint for debugging to get all
the qiskit package versions. This is undesirable because the
qiskit-ibmq-provider usage of the __qiskit_version__ attribute happens
on import, this means that we'll be unconditionally importing the entire
set of qiskit packages even if only terra and the qiskit provider are
used. This has the negative effect of both significantly increasing
import time (as things like aqua have very heavy dependencies and a slow
import tree) but also by potentially causing undesireable deprecation
warnings to be emitted if there are things in the import path that are
deprecated (like aqua). To fix this issue this commit switches away from
using the __qiskit_version__ attribute and queries the qiskit
metapackage version directly, and if it's not available then the ibmq
provider and terra versions are set in the header. This avoids the
import related issues from using __qiskit_version__ while not changing
the functionality (because the other non-terra requirements in the
header should have no impact on the api behavior).

The client version used in the header should honestly be independent of
anything outside of the qiskit-ibmq-provider as it's the only
interaction point between the iqx api and qiskit (and things I believe
this was added for like the basis gate change can be handled by the
minimum version of terra in the requirements). Doing it this way just adds
an unnecessary dependency and complexity to version negotiation. But it's
too late to change the iqx api design at this point.

* update header format

Co-authored-by: jessieyu <jessieyu@us.ibm.com>
(cherry picked from commit f5c3b5003e2c71c3befba7f410cbcba4008f6771)

* Backported from: Make AnalysisResult.fit optional (#917)

* update version

Co-authored-by: Matt Riedemann <mriedem@ibm.com>
Co-authored-by: Matthew Treinish <mtreinish@kortar.org>