Remove deprecated ResultQuality values (#919)
* Remove deprecated ResultQuality values

The resultsdb service has deprecated the Human/Computer
analysis result `quality` values. They are replaced with
just `Good` and `Bad` values to be used with the `verified`
flag on an analysis result.

This builds on #909 and #915 to remove the deprecated
enum quality values, gets the skipped test working again,
and adds an upgrade release note. Since the experiment
service in Qiskit is still considered beta there is no
deprecation period in Qiskit itself though the service/API
continues to honor and translate the old quality values
for now.

Closes #902

* Fix lint errors

* Support updating AnalysisResult.verified

This makes `test_update_analysis_result` pass now that
we have removed the `HUMAN_BAD` quality value.

* Rename NO_INFORMATION to UNKNOWN

Per review feedback this is getting renamed to
be a single word. The underlying value does not
change.