Fix test_upload_update_plot_data by using synchronous upload
Version 0.34.9 of the experiments service REST API changed the
default behavior on the `POST /experiments/:uuid/plots` and
`PUT /experiments/:uuid/plots/:name` APIs to upload the plot file
to backend (object) storage asynchronously.

That broke the `test_upload_update_plot_data` test which creates
a plot, updates it and then immediately gets the plot back and
compares the updated plot against the original plot. With the async
upload change on the server side the test intermittently fails due
to a race between the PUT and GET requests for the plot.

This change fixes that by making the plot update PUT request
synchronous by passing the `x-sync-upload: "True"` request header.
This way the client can be sure the plot was updated by the time
the PUT response is returned and before GET'ing the updated plot.

The default behavior on the Qiskit client side remains the same,
i.e. asynchronous plot uploads to backend storage, since in most
cases the client likely does not need to wait for the plot to be
uploaded on the backend.

Closes #992