Fix test_upload_update_plot_data by using synchronous upload (#998)
* Fix test_upload_update_plot_data by using synchronous upload

Version 0.34.9 of the experiments service REST API changed the
default behavior on the `POST /experiments/:uuid/plots` and
`PUT /experiments/:uuid/plots/:name` APIs to upload the plot file
to backend (object) storage asynchronously.

That broke the `test_upload_update_plot_data` test which creates
a plot, updates it and then immediately gets the plot back and
compares the updated plot against the original plot. With the async
upload change on the server side the test intermittently fails due
to a race between the PUT and GET requests for the plot.

This change fixes that by making the plot update PUT request
synchronous by passing the `x-sync-upload: "True"` request header.
This way the client can be sure the plot was updated by the time
the PUT response is returned and before GET'ing the updated plot.

The default behavior on the Qiskit client side remains the same,
i.e. asynchronous plot uploads to backend storage, since in most
cases the client likely does not need to wait for the plot to be
uploaded on the backend.

Closes #992

* Change plot updates to be synchronous by default

Based on feedback in PR #998 plot updates will now
be uploaded to the backend storage synchronously
by default. This is so the client can be sure the plot
upload completed successfully.

A follow up change will add the same behavior to the
plot creation code.

* Make plot upload synchronous

The `POST /experiments/:uuid/plots` API now uploads the
plot to backend storage asynchronously. This change makes
the plot creation/upload synchronous by default (per the client)
like before and also the same default behavior as the plot update.

* Add release note for `sync_upload` keyword

* wording

* fix lint

Co-authored-by: Jessie Yu <jessieyu@us.ibm.com>