Remove invalid type hint from least_busy method
This commit removes an invalid type hint from the least_busy method.
This method was incorrectly listing the BaseBackend class as a valid
input and output type from the method. This has not been a supported
type since #817. Additionally, the class has been deprecated in
qiskit-terra for ~1 year now and was recently removed from the main
development branch. The ibmq provider should not be importing using
this legacy type, especially for a type hint as it is not used
anywhere.