Fix version string parsing of terra
In the upcoming terra release there will be a release candidate tagged
prior to the final release. However changing the version string for the
package is blocked on the ibmq provider right now because it is trying
to parse the version and is assuming there will be no prelease suffix on
the version string (see Qiskit/qiskit-terra#8200 for the details). This
commit fixes this version parsing to use the regex from the
pypa/packaging project which handles all the PEP440 package versioning
include pre-release suffixes. This will enable terra to release an
0.21.0rc1 tag without breaking the ibmq provider.