Revert "fix python requirement versions (#1159)"
This reverts commit 2ad2352fbba0962f33017339965e93c86e3b84b3. The
changes made to the requirements list in that commit were not valid and
actively prevent users from installing the qiskit-ibmq-provider with any
version of qiskit-terra other than the 0.23.x release series. Similarly
the caps set on every dependency are overly restrictive and will cause
compatibility issues with other packages, especially if there are shared
dependencies between multiple packages. This commit reverts these
changes to prevent issues and this needs to be released as part as
0.20.2 to unblock a future qiskit metapackage release.