Make deploy documentation script executable
In the recently merged #1160 a documentation build and publishing
pipeline was added to gha so that we're able to publish the
documentation for the ibmq provider to a standalone page. However, one
small oversight in that PR is that the deployment script was not set as
executable so when the action tries to directly execute the script it
fails. This commit fixes the permissions of the file in the repo so that
the script can be run directly.