Remove scikit-quant test usage and unpin numpy (#1171)
The numpy requirement was pinned to avoid a CI issue caused by an
incompatibility with the scikit-quant package (see:
https://github.com/scikit-quant/scikit-quant/issues/24). However, the
requirements.txt file has wider reach than just the test environment and
it is actively causing conflicts with projects using numpy and
qiskit-ibmq-provider as dependencies. The normal path to fix this is to
pin the numpy version in the environments that are using both
scikit-quant and numpy. Since the testing with scikit-quant isn't
critical instead of pinning numpy in CI environments this commit opts to
just remove the scikit-quant usage instead.

See Qiskit/qiskit-ibm-runtime#845 for a similar change made there.

Co-authored-by: Kevin Tian <kevin.tian@ibm.com>